Stabilize DQN further...
by decreasing learning rate and getting rid of last layer biases. Last
layer biases take the whole gradient on themselfs causing the DQN
to become less and less dependent on the state so it outputs always
the same values.

Additionally, add logs when the DQN outputs super simillar values for 2 different
states to spot whenever DQN stops to depend on states and tries to generalize
everything just based on the action.